Move UniqueHashDictionary to Server
