--- src/MagicOnion.Client.Unity/Assets/Scripts/MagicOnion/MagicOnion.Client/MagicOnion.Shared/Utils/UniqueHashDictionary.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f46f5d25ac409534981ec3509fb8859e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
